requirements.txt
lxml==5.1.0
python-quickbooks==0.9.7
intuitlib==1.3.2
flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
imapclient==3.0.1
apscheduler==3.10.4
requests==2.31.0
